Move fixtures to separate conftest.py
Makes it possible to reuse fixtures across test files
